Refactor content actions into separate files.
