+++ src/routes/content/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+exports.store = require('./store');
+exports.bulk = require('./bulk');
+exports.retrieve = require('./retrieve');
+exports.remove = require('./remove');
